This commit adds two new methods to openssl.ssl.ctx objects.
- `ctx:setCertificteFromFile` calls `SSL_CTX_use_certificate_chain_file`
to add a certificate chain from a pem encoded file specified by the string argument path.

- `ctx:setPrivateKeyFromFile` calls `SSL_CTX_use_private_key_file`
to add a private key from a PEM or ASN1 encoded file using the string argument path
and filetype integer flag argument. The filetype is optional and will default to PEM if not
specified.

- `openssl.filetypes` is a new table in the openssl module which contains the
two filetypes used by `setPrivateKeyFromFile`. The `.PEM` field is the value of `SSL_FILETYPE_PEM` and
the `.ASN1` field is the value of `SSL_FILETYPE_ASN1`.